+++ src/Composer/Plugin/MakeBomCommand.php
@@ -24,15 +24,17 @@ declare(strict_types=1);
 namespace CycloneDX\Composer\Plugin;
 
 use Composer\Command\BaseCommand;
-use Composer\Composer;
 use CycloneDX\Composer\Factories\BomFactory;
 use CycloneDX\Composer\Factories\ComponentFactory;
 use CycloneDX\Composer\Factories\LicenseFactory;
 use CycloneDX\Composer\Factories\SpecFactory;
 use CycloneDX\Composer\Locker;
 use CycloneDX\Composer\Plugin\Exceptions\ValueError;
+use CycloneDX\Core\Models\Bom;
 use CycloneDX\Core\Serialize\SerializerInterface;
 use CycloneDX\Core\Spdx\License as SpdxLicenseValidator;
+use CycloneDX\Core\Validation\ValidationError;
+use CycloneDX\Core\Validation\ValidatorInterface;
 use RuntimeException;
 use Symfony\Component\Console\Formatter\OutputFormatter;
 use Symfony\Component\Console\Input\InputInterface;
@@ -62,8 +64,10 @@ class MakeBomCommand extends BaseCommand
     /**
      * @throws RuntimeException
      */
-    private function compat_getComposer(): Composer
+    private function getLocker(OutputInterface $output): ?\Composer\Package\Locker
     {
+        $output->writeln('<info>Gathering lockfile</info>', OutputInterface::VERBOSITY_VERY_VERBOSE);
+
         try {
             $composer = $this->getComposer();
         } catch (\Exception $exception) {
@@ -73,7 +77,11 @@ class MakeBomCommand extends BaseCommand
             throw new UnexpectedValueException('Composer is null');
         }
 
-        return $composer;
+        $locker = $composer->getLocker();
+
+        return $locker->isLocked()
+            ? $locker
+            : null;
     }
 
     /**
@@ -81,8 +89,8 @@ class MakeBomCommand extends BaseCommand
      */
     protected function execute(InputInterface $input, OutputInterface $output): int
     {
-        $locker = $this->compat_getComposer()->getLocker();
-        if (false === $locker->isLocked()) {
+        $locker = $this->getLocker($output);
+        if (null === $locker) {
             $output->writeln('<error>Lockfile does not exist</error>');
 
             return self::INVALID;
@@ -95,19 +103,29 @@ class MakeBomCommand extends BaseCommand
 
             return self::INVALID;
         }
+        $output->writeln(
+            __METHOD__.' Options: '.print_r($options, true),
+            OutputInterface::VERBOSITY_DEBUG | OutputInterface::OUTPUT_RAW
+        );
 
-        return $this->makeAndWrite(new Locker($locker), $options, $output)
+        $bom = $this->makeBom(new Locker($locker), $options, $output);
+
+        return $this->writeBom($bom, $options, $output)
             ? self::SUCCESS
             : self::FAILURE;
     }
 
     /**
-     * @psalm-suppress MissingThrowsDocblock - Exceptions are handled by caller
-     * @psalm-suppress RedundantConditionGivenDocblockType - Some IDEs need the assert as a help
+     * @throws RuntimeException
+     * @throws \DomainException
      */
-    private function makeAndWrite(Locker $locker, MakeBomCommandOptions $options, OutputInterface $output): bool
+    private function makeBom(Locker $locker, MakeBomCommandOptions $options, OutputInterface $output): Bom
     {
-        $output->writeln('<info>Generating BOM from lockfile</info>', OutputInterface::VERBOSITY_VERBOSE);
+        $output->writeln(
+            '<info>Generating BOM from lockfile</info>',
+            OutputInterface::VERBOSITY_VERBOSE
+        );
+
         $bom = (new BomFactory(
             $options->excludeDev, $options->excludePlugins,
             new ComponentFactory(
@@ -117,37 +135,124 @@ class MakeBomCommand extends BaseCommand
             )
         ))->makeFromLocker($locker);
 
-        $spec = (new SpecFactory())->make($options->specVersion);
-        $bomWriter = new $options->bomWriterClass($spec);
-        \assert($bomWriter instanceof SerializerInterface);
-
         $output->writeln(
-            '<info>Serializing BOM: '.OutputFormatter::escape($options->bomFormat).' '.OutputFormatter::escape(
-                $options->specVersion
-            ).'</info>',
-            OutputInterface::VERBOSITY_VERBOSE
+            'Bom: '.print_r($bom, true),
+            OutputInterface::VERBOSITY_DEBUG | OutputInterface::OUTPUT_RAW
         );
-        $bomContents = $bomWriter->serialize($bom);
 
-        if (MakeBomCommandOptions::OUTPUT_FILE_STDOUT === $options->outputFile) {
-            $output->writeln('<info>Writing output to STDOUT</info>', OutputInterface::VERBOSITY_VERBOSE);
-            // don't use `$output->writeln()`, so to support `-q` cli param and still have the desired output.
-            return false !== fwrite(\STDOUT, $bomContents);
+        return $bom;
+    }
+
+    /**
+     * @throws \InvalidArgumentException if SPEC version is unknown
+     */
+    private function writeBom(Bom $bom, MakeBomCommandOptions $options, OutputInterface $output): bool
+    {
+        $spec = (new SpecFactory())->make($options->specVersion);
+
+        $bomString = $this->makeBomString(new $options->bomWriterClass($spec), $bom, $output);
+
+        $isValid = $this->validateBomString(
+            $bomString,
+            null === $options->bomValidatorClass ? null : new $options->bomValidatorClass($spec),
+            $output
+        );
+        if (false === $isValid) {
+            return false;
         }
 
-        $output->writeln('<info>Writing output to: '.OutputFormatter::escape($options->outputFile).'</info>', OutputInterface::VERBOSITY_VERBOSE);
-        $written = file_put_contents($options->outputFile, $bomContents);
+        if (MakeBomCommandOptions::OUTPUT_FILE_STDOUT === $options->outputFile) {
+            $output->writeln(
+                '<info>Writing output to STDOUT</info>',
+                OutputInterface::VERBOSITY_VERBOSE
+            );
+
+            // don't use `$output->writeln()`, so to support `-q` cli param and still have the desired output.
+            return false !== fwrite(\STDOUT, $bomString);
+        }
+
+        $output->writeln(
+            '<info>Writing output to: '.
+            OutputFormatter::escape($options->outputFile).
+            '</info>',
+            OutputInterface::VERBOSITY_VERBOSE
+        );
+        $written = file_put_contents($options->outputFile, $bomString);
         if (false === $written) {
-            $output->writeln('<error>Failed writing to file: '.OutputFormatter::escape($options->outputFile).'</error>');
+            $output->writeln(
+                '<error>Failed writing to file: '.
+                OutputFormatter::escape($options->outputFile).
+                '</error>'
+            );
 
             return false;
         }
         $output->writeln(
-            '<info>Wrote '.OutputFormatter::escape($options->bomFormat).' '.OutputFormatter::escape(
-                $options->specVersion
-            ).' to: '.OutputFormatter::escape($options->outputFile).'</info>'
+            '<info>Wrote '.
+            OutputFormatter::escape($options->bomFormat).' '.
+            OutputFormatter::escape($options->specVersion).
+            ' to: '.OutputFormatter::escape($options->outputFile).
+            '</info>'
         );
 
         return true;
+    }
+
+    private function makeBomString(SerializerInterface $bomWriter, Bom $bom, OutputInterface $output): string
+    {
+        $output->writeln(
+            '<info>Serializing BOM with '.
+            OutputFormatter::escape(\get_class($bomWriter)).
+            ' for '.OutputFormatter::escape($bomWriter->getSpec()->getVersion()).
+            '</info>',
+            OutputInterface::VERBOSITY_VERY_VERBOSE
+        );
+
+        return $bomWriter->serialize($bom);
+    }
+
+    private function validateBomString(string $bom, ?ValidatorInterface $validator, OutputInterface $output): ?bool
+    {
+        if (null === $validator) {
+            $output->writeln(
+                '<info>Skipping output validation</info>',
+                OutputInterface::VERBOSITY_VERY_VERBOSE
+            );
+
+            return null;
+        }
+
+        $output->writeln(
+            '<info>Validating output with '.
+            OutputFormatter::escape(\get_class($validator)).
+            ' for '.OutputFormatter::escape($validator->getSpec()->getVersion()).
+            '</info>',
+            OutputInterface::VERBOSITY_VERY_VERBOSE
+        );
+
+        try {
+            $validationError = $validator->validateString($bom);
+        } catch (\Exception $exception) {
+            $validationError = ValidationError::fromThrowable($exception);
+        }
+
+        if (null === $validationError) {
+            return true;
+        }
+
+        $output->writeln(
+            print_r($validationError, true),
+            OutputInterface::VERBOSITY_DEBUG | OutputInterface::OUTPUT_RAW
+        );
+
+        $output->writeln(
+            [
+                '<error>Failed to generate valid output.</error>',
+                '<warning>Please report the issue and provide the "composer.lock" file of the current project to:</warning>',
+                '<warning>https://github.com/CycloneDX/cyclonedx-php-composer/issues/new</warning>',
+            ]
+        );
+
+        return false;
     }
 }