+++ tests/Core/Serialize/DOM/Normalizers/ComponentRepositoryNormalizerTest.php
@@ -35,6 +35,7 @@ use PHPUnit\Framework\TestCase;
 /**
  * @covers \CycloneDX\Core\Serialize\DOM\Normalizers\ComponentRepositoryNormalizer
  * @covers \CycloneDX\Core\Serialize\DOM\AbstractNormalizer
+ * @covers \CycloneDX\Core\Helpers\SimpleDomTrait
  */
 class ComponentRepositoryNormalizerTest extends TestCase
 {