+++ tests/Core/Serialize/DOM/Normalizers/DisjunctiveLicenseNormalizerTest.php
@@ -36,6 +36,7 @@ use PHPUnit\Framework\TestCase;
 /**
  * @covers \CycloneDX\Core\Serialize\DOM\Normalizers\DisjunctiveLicenseNormalizer
  * @covers \CycloneDX\Core\Serialize\DOM\AbstractNormalizer
+ * @covers \CycloneDX\Core\Helpers\SimpleDomTrait
  */
 class DisjunctiveLicenseNormalizerTest extends TestCase
 {